requirements.txt
isort==5.9.3
peewee==3.14.8
pyserial==3.5
